FTP user cmd response handling fix. GArray::AddTo fix. Debugger fixes.

+++ include/common/IFtp.h
@@ -145,7 +145,7 @@ protected:
 	// void *FactoryParam;
 
 	// State
-	int RestorePos;
+	int64 RestorePos;
 	bool AbortTransfer;
 	
 	bool SetupData(bool Binary);
